change declaration name for testing. define overload factory func
